Changed the \includearticle command.
Problem: The current implementation forced the pdf file to be put in
the root folder, and the file to be named identically to the citekey.
This was implemented in the includearticle macro. The includearticle
macro also forced centering withou alignment of the output from
\fullcite.  This is preferable for text in general on otherwise blank
pages, but looked rather ill formatted for a citation.

This commit simply separates the argument of file path and citekey
to two separate arguments in \includearticle. The optional pages
parameter is also added to the README.md. This was undocumented before
and is good. Often, papers have a cover page with metadata content from
the publisher. To alleviate the output from \fullcite, this commit puts
the citation in a flushleft environment for alignment. This means it is
still centered, but left aligned.